📝 Improve getting started instructions
Use a numbered system for the "get started" started section. The previous README wasn't as clear as it could be with its instructions. This way, a user can more easily follow the four steps to get started.

Also added some Foundation 6 elements such as mixin commenting to not include any unused Foundation CSS.